Cleaned up more code. Changed some floats to doubles in drrt_data_structures.h. Added Problem class for RRT problems. 	These contain all the information I needed to run RRTX() and the algorithm in general. 	Executable just needs to include drrt.h 	Executable should implement a main, an RRT function, and a distance function (see test.cpp).

+++ include/DRRT/drrt_data_structures.h
@@ -28,7 +28,7 @@ public:
                         // is theta in particular a Dubin's system is used
 
     // Stuff for sampling functions
-    float pGoal;  // the probabality that the goal is sampled
+    double pGoal;  // the probabality that the goal is sampled
     std::string randNode; // the sampling function to use (takes a CSpace)
 
     std::shared_ptr<KDTreeNode> goalNode;  // the goal node
@@ -38,28 +38,28 @@ public:
     int itsUntilSample;    // a count down to sample a particular point
     Eigen::VectorXd itsSamplePoint;   // sample this when itsUntilSample == 0
     Eigen::VectorXd timeSamplePoint;  // sample this when waitTime has passed
-    float waitTime;                   // time to wait in seconds
+    double waitTime;                   // time to wait in seconds
     u_int64_t startTimeNs;            // time this started
-    float timeElapsed; // elapsed time since started ( where time spent saving
+    double timeElapsed; // elapsed time since started ( where time spent saving
                        // experimental data has been removed )
 
     //Obstacle* obstacleToRemove;    // an obstacle to remove
 
-    float robotRadius;       // robot radius
-    float robotVelocity;     // robot velocity (used for Dubins w/o time)
+    double robotRadius;       // robot radius
+    double robotVelocity;     // robot velocity (used for Dubins w/o time)
 
-    float dubinsMinVelocity; // min velocity of Dubin's car (for dubins + time)
-    float dubinsMaxVelocity; // max velocity of Dubin's car (for dubins + time)
+    double dubinsMinVelocity; // min velocity of Dubin's car (for dubins + time)
+    double dubinsMaxVelocity; // max velocity of Dubin's car (for dubins + time)
 
     // This jlist must "hold" MatrixXd's ?
     // So use the JListNode->node->position when using this stack
     std::shared_ptr<JList> sampleStack; // points to sample in the future
 
-    float hypervolume;          // hypervolume of the space
-    float delta;                // RRT parameter delta
-    float minTurningRadius;     // min turning radius e.g. for Dubin's car
+    double hypervolume;          // hypervolume of the space
+    double delta;                // RRT parameter delta
+    double minTurningRadius;     // min turning radius e.g. for Dubin's car
 
-    float warmupTime;   // the amount of warm up time allowed (obstacles are
+    double warmupTime;   // the amount of warm up time allowed (obstacles are
                         // ignored for warm up time)
     bool inWarmupTime;  // true if we are in the warm up time
 
@@ -85,7 +85,7 @@ typedef struct Queue{
     std::shared_ptr<CSpace> S;
     BinaryHeap* Q;      // normal queue (sorted based on cost from goal)
     std::shared_ptr<JList> OS;          // obstacle successor stack
-    float changeThresh; // threshold of local changes that we care about
+    double changeThresh; // threshold of local changes that we care about
 
 } Queue;
 
@@ -160,7 +160,7 @@ typedef struct RobotData{
 
     Eigen::MatrixXd robotLocalPath; // this holds the path between robotPose
                              // and nextRobotPose (not including the former)
-    float numLocalMovePoints;   // the number of points in robotLocalPath
+    double numLocalMovePoints;   // the number of points in robotLocalPath
 
     std::shared_ptr<Edge> robotEdge; // this is the edge that contains the
                                      // trajectory that the
@@ -172,22 +172,29 @@ typedef struct RobotData{
     // is used at a time.
     // Which one is used depends on if time is explicitely part of
     // the state space.
-    float distAlongRobotEdge; // the current distance that the robot
+    double distAlongRobotEdge; // the current distance that the robot
                               // "will be" along robotEdge (next time slice)
 
-    float timeAlongRobotEdge; // the current time that the robot "will be"
+    double timeAlongRobotEdge; // the current time that the robot "will be"
                               // along robotEdge (i.e. next time slice)
 
     // Constructor
-    RobotData(Eigen::VectorXd rP, std::shared_ptr<KDTreeNode> nMT,
-              int maxPathNodes) :
-        robotPose(rP), nextRobotPose(rP), nextMoveTarget(nMT),
-        distanceFromNextRobotPoseToNextMoveTarget(0.0), moving(false),
-        currentMoveInvalid(false), numRobotMovePoints(1),
-        /*Original Julia code does not have an
-         * initial value for numLocalMovePoints*/
-        numLocalMovePoints(1), robotEdgeUsed(false),
-        distAlongRobotEdge(0.0), timeAlongRobotEdge(0.0)
+    RobotData(Eigen::VectorXd rP,
+              std::shared_ptr<KDTreeNode> nMT,
+              int maxPathNodes)
+        : robotPose(rP),
+          nextRobotPose(rP),
+          nextMoveTarget(nMT),
+          distanceFromNextRobotPoseToNextMoveTarget(0.0),
+          moving(false),
+          currentMoveInvalid(false),
+          numRobotMovePoints(1),
+          /*Original Julia code does not have an
+           * initial value for numLocalMovePoints*/
+          numLocalMovePoints(1),
+          robotEdgeUsed(false),
+          distAlongRobotEdge(0.0),
+          timeAlongRobotEdge(0.0)
     {
         robotLocalPath.resize(maxPathNodes,4);
         robotMovePath.resize(maxPathNodes,4);
